Workaround vrc GetBonePosition NullReferenceException bug
This introduces a sacrificial prefab to perform bone location queries.
In the event of an uncaught exception propagated up from
VRCPlayerApi.GetBonePosition or GetBoneRotation, the UdonBehaviour on
this prefab will die, but this will then be detected and the prefab will
be re-instantiated to return to normal operation.

Additionally, the last player queried will not be queried again for one
second, to attempt to avoid the issue occurring again immediately.