Make global tracker search more reliable and efficient
The existing code used U#'s GetUdonSharpComponentInChildren to locate
the scene-global tracker object. However, this could fail in some
situations (probably when there are other broken U# components in the
scene). It's also very inefficient as U# would load the proxy for all
objects before doing a typecheck.

Instead, we now search for the UdonBehaviour directly, and check the
programSource parameter for the script of interest. We then convert to
the proxy object only when we find the component we were looking for.